Improve retrieval of available language mappings - dynamic including custom defined languages rather than using the constants.
Currently strips off the "text/" namespace at the start of most of them,
but in doing so maintains back compatibility. This means that for
example "application/javascript" and "text/javascript" if they were both
written might resolve to the same thing. As it stands, everything uses
the text/ prefix and so it's not a problem.